Refactor: split wordpress.js into domain-specific lib files, remove Apollo Client, restore wp-search.json path
