+++ package.json
@@ -15,7 +15,6 @@
     "prepare": "husky"
   },
   "dependencies": {
-    "@apollo/client": "^4.1.3",
     "@cloudinary/url-gen": "^1.22.0",
     "clsx": "^2.1.1",
     "date-fns": "^4.1.0",
@@ -23,7 +22,6 @@
     "dotenv": "^17.2.2",
     "fabric": "^6.7.1",
     "fuse.js": "^7.1.0",
-    "graphql": "^16.12.0",
     "he": "^1.2.0",
     "loader-utils": "^3.3.1",
     "next": "^16.1.6",
