+++ src/components/Metadata/Metadata.js
@@ -1,7 +1,7 @@
 import Link from 'next/link';
 import clsx from 'clsx';
 
-import { categoryPathBySlug } from 'lib/wordpress';
+import { categoryPathBySlug } from 'lib/categories';
 import { formatDate } from 'lib/datetime';
 
 import styles from './Metadata.module.scss';
