+++ src/components/Nav/Nav.js
@@ -5,7 +5,7 @@ import Link from 'next/link';
 import { FaSearch } from 'react-icons/fa';
 
 import useSearch, { SEARCH_STATE_LOADED } from 'hooks/use-search';
-import { postPathBySlug } from 'lib/wordpress';
+import { postPathBySlug } from 'lib/posts';
 
 import Section from 'components/Section';
 import Container from 'components/Container';
