--- src/data/pages.js
@@ -1,143 +1,0 @@
-import { gql } from '@apollo/client';
-
-export const QUERY_ALL_PAGES = gql`
-  {
-    pages(first: 10000) {
-      edges {
-        node {
-          children {
-            edges {
-              node {
-                id
-                slug
-                uri
-                ... on Page {
-                  id
-                  title
-                }
-              }
-            }
-          }
-          content
-          featuredImage {
-            node {
-              altText
-              caption
-              id
-              sizes
-              sourceUrl
-              srcSet
-            }
-          }
-          id
-          menuOrder
-          modified
-          page {
-            excludeFromPathGeneration
-          }
-          parent {
-            node {
-              id
-              slug
-              uri
-              ... on Page {
-                title
-              }
-            }
-          }
-          slug
-          title
-          uri
-        }
-      }
-    }
-  }
-`;
-
-export const QUERY_PAGE_BY_URI = gql`
-  query PageByUri($uri: ID!) {
-    page(id: $uri, idType: URI) {
-      children {
-        edges {
-          node {
-            id
-            slug
-            uri
-            ... on Page {
-              id
-              title
-            }
-          }
-        }
-      }
-      content
-      featuredImage {
-        node {
-          altText
-          caption
-          id
-          sizes
-          sourceUrl
-          srcSet
-        }
-      }
-      id
-      menuOrder
-      modified
-      parent {
-        node {
-          id
-          slug
-          uri
-          ... on Page {
-            title
-          }
-        }
-      }
-      slug
-      title
-      uri
-    }
-  }
-`;
-
-export const QUERY_PAGE_SEO_BY_URI = gql`
-  query PageSEOByUri($uri: ID!) {
-    page(id: $uri, idType: URI) {
-      id
-      seo {
-        canonical
-        metaDesc
-        metaRobotsNofollow
-        metaRobotsNoindex
-        opengraphAuthor
-        opengraphDescription
-        opengraphModifiedTime
-        opengraphPublishedTime
-        opengraphPublisher
-        opengraphTitle
-        opengraphType
-        readingTime
-        title
-        twitterDescription
-        twitterTitle
-        twitterImage {
-          altText
-          sourceUrl
-          mediaDetails {
-            width
-            height
-          }
-        }
-        opengraphImage {
-          altText
-          sourceUrl
-          mediaDetails {
-            height
-            width
-          }
-        }
-      }
-    }
-  }
-`;
