--- src/data/users.js
@@ -1,55 +1,0 @@
-import { gql } from '@apollo/client';
-
-export const QUERY_ALL_USERS = gql`
-  {
-    users(first: 10000) {
-      edges {
-        node {
-          avatar {
-            height
-            width
-            url
-          }
-          description
-          id
-          name
-          roles {
-            nodes {
-              name
-            }
-          }
-          slug
-        }
-      }
-    }
-  }
-`;
-
-export const QUERY_ALL_USERS_SEO = gql`
-  {
-    users(first: 10000) {
-      edges {
-        node {
-          id
-          seo {
-            metaDesc
-            metaRobotsNofollow
-            metaRobotsNoindex
-            title
-            social {
-              youTube
-              wikipedia
-              twitter
-              soundCloud
-              pinterest
-              mySpace
-              linkedIn
-              instagram
-              facebook
-            }
-          }
-        }
-      }
-    }
-  }
-`;
