+++ src/lib/pages.js
@@ -1,204 +1,246 @@
-import { getApolloClient } from 'lib/apollo-client';
-
-import { QUERY_ALL_PAGES, QUERY_PAGE_BY_URI, QUERY_PAGE_SEO_BY_URI } from 'data/pages';
-
-/**
- * pagePathBySlug
- */
-
-export function pagePathBySlug(slug) {
-  return `/${slug}`;
-}
-
-/**
- * getPageByUri
- */
+import { fetchAPI } from 'lib/api';
 
 export async function getPageByUri(uri) {
-  const apolloClient = getApolloClient();
-  const apiHost = new URL(process.env.WORDPRESS_GRAPHQL_ENDPOINT).host;
+  const query = `
+    query PageByUri($uri: ID!) {
+      page(id: $uri, idType: URI) {
+        children {
+          edges {
+            node {
+              id
+              slug
+              uri
+              ... on Page {
+                id
+                title
+              }
+            }
+          }
+        }
+        content
+        featuredImage {
+          node {
+            altText
+            caption
+            id
+            sizes
+            sourceUrl
+            srcSet
+          }
+        }
+        id
+        menuOrder
+        modified
+        parent {
+          node {
+            id
+            slug
+            uri
+            ... on Page {
+              title
+            }
+          }
+        }
+        slug
+        title
+        uri
+      }
+    }
+  `;
 
-  let pageData;
-  let seoData;
+  const data = await fetchAPI(query, { uri });
 
-  try {
-    pageData = await apolloClient.query({
-      query: QUERY_PAGE_BY_URI,
-      variables: {
-        uri,
-      },
-    });
-  } catch (e) {
-    console.log(`[pages][getPageByUri] Failed to query page data: ${e.message}`);
-    throw e;
+  if (!data?.page) {
+    return { page: undefined };
   }
 
-  if (!pageData?.data.page) {
-    return {
-      page: undefined,
-    };
+  const page = { ...data.page };
+
+  if (page.featuredImage) {
+    page.featuredImage = page.featuredImage.node;
   }
 
-  const page = [pageData?.data.page].map(mapPageData)[0];
+  if (page.parent?.node) {
+    page.parent = page.parent.node;
+  }
 
-  // If the SEO plugin is enabled, look up the data
-  // and apply it to the default settings
+  if (page.children?.edges) {
+    page.children = page.children.edges.map(({ node }) => node);
+  }
 
   if (process.env.WORDPRESS_PLUGIN_SEO === true) {
+    const seoQuery = `
+      query PageSEOByUri($uri: ID!) {
+        page(id: $uri, idType: URI) {
+          id
+          seo {
+            canonical
+            metaDesc
+            metaRobotsNofollow
+            metaRobotsNoindex
+            opengraphAuthor
+            opengraphDescription
+            opengraphModifiedTime
+            opengraphPublishedTime
+            opengraphPublisher
+            opengraphTitle
+            opengraphType
+            readingTime
+            title
+            twitterDescription
+            twitterTitle
+            twitterImage {
+              altText
+              sourceUrl
+              mediaDetails {
+                width
+                height
+              }
+            }
+            opengraphImage {
+              altText
+              sourceUrl
+              mediaDetails {
+                height
+                width
+              }
+            }
+          }
+        }
+      }
+    `;
+
     try {
-      seoData = await apolloClient.query({
-        query: QUERY_PAGE_SEO_BY_URI,
-        variables: {
-          uri,
-        },
-      });
+      const seoData = await fetchAPI(seoQuery, { uri });
+      if (seoData?.page?.seo) {
+        const { seo = {} } = seoData.page;
+        page.metaTitle = seo.title;
+        page.description = seo.metaDesc;
+        page.readingTime = seo.readingTime;
+        page.canonical = seo.canonical;
+        page.og = {
+          author: seo.opengraphAuthor,
+          description: seo.opengraphDescription,
+          image: seo.opengraphImage,
+          modifiedTime: seo.opengraphModifiedTime,
+          publishedTime: seo.opengraphPublishedTime,
+          publisher: seo.opengraphPublisher,
+          title: seo.opengraphTitle,
+          type: seo.opengraphType,
+        };
+        page.robots = {
+          nofollow: seo.metaRobotsNofollow,
+          noindex: seo.metaRobotsNoindex,
+        };
+        page.twitter = {
+          description: seo.twitterDescription,
+          image: seo.twitterImage,
+          title: seo.twitterTitle,
+        };
+      }
     } catch (e) {
       console.log(`[pages][getPageByUri] Failed to query SEO plugin: ${e.message}`);
-      console.log('Is the SEO Plugin installed? If not, disable WORDPRESS_PLUGIN_SEO in next.config.js.');
-      throw e;
-    }
-
-    if (seoData?.data?.page) {
-      const { seo = {} } = seoData.data.page;
-
-      page.metaTitle = seo.title;
-      page.description = seo.metaDesc;
-      page.readingTime = seo.readingTime;
-
-      // The SEO plugin by default includes a canonical link, but we don't want to use that
-      // because it includes the WordPress host, not the site host. We manage the canonical
-      // link along with the other metadata, but explicitly check if there's a custom one
-      // in here by looking for the API's host in the provided canonical link
-
-      if (seo.canonical && !seo.canonical.includes(apiHost)) {
-        page.canonical = seo.canonical;
-      }
-
-      page.og = {
-        author: seo.opengraphAuthor,
-        description: seo.opengraphDescription,
-        image: seo.opengraphImage,
-        modifiedTime: seo.opengraphModifiedTime,
-        publishedTime: seo.opengraphPublishedTime,
-        publisher: seo.opengraphPublisher,
-        title: seo.opengraphTitle,
-        type: seo.opengraphType,
-      };
-
-      page.robots = {
-        nofollow: seo.metaRobotsNofollow,
-        noindex: seo.metaRobotsNoindex,
-      };
-
-      page.twitter = {
-        description: seo.twitterDescription,
-        image: seo.twitterImage,
-        title: seo.twitterTitle,
-      };
     }
   }
 
-  return {
-    page,
-  };
+  return { page };
 }
-
-/**
- * getAllPages
- */
 
 export async function getAllPages() {
-  const apolloClient = getApolloClient();
+  const query = `
+    {
+      pages(first: 10000) {
+        edges {
+          node {
+            children {
+              edges {
+                node {
+                  id
+                  slug
+                  uri
+                  ... on Page {
+                    id
+                    title
+                  }
+                }
+              }
+            }
+            content
+            featuredImage {
+              node {
+                altText
+                caption
+                id
+                sizes
+                sourceUrl
+                srcSet
+              }
+            }
+            id
+            menuOrder
+            modified
+            page {
+              excludeFromPathGeneration
+            }
+            parent {
+              node {
+                id
+                slug
+                uri
+                ... on Page {
+                  title
+                }
+              }
+            }
+            slug
+            title
+            uri
+          }
+        }
+      }
+    }
+  `;
 
-  const data = await apolloClient.query({
-    query: QUERY_ALL_PAGES,
-  });
+  const data = await fetchAPI(query);
+  const pages =
+    data?.pages?.edges?.map(({ node }) => {
+      const page = { ...node };
 
-  const pages = data?.data.pages.edges.map(({ node = {} }) => node).map(mapPageData);
+      if (page.featuredImage?.node) {
+        page.featuredImage = page.featuredImage.node;
+      }
 
-  return {
-    pages,
-  };
+      if (page.parent?.node) {
+        page.parent = page.parent.node;
+      }
+
+      if (page.children?.edges) {
+        page.children = page.children.edges.map(({ node: child }) => child);
+      }
+
+      return page;
+    }) || [];
+
+  return { pages };
 }
 
-/**
- * getTopLevelPages
- */
+export function getBreadcrumbsByUri(uri, pages = []) {
+  if (!uri || uri === '/') return [];
 
-export async function getTopLevelPages() {
-  const { pages } = await getAllPages();
-
-  const navPages = pages.filter(({ parent }) => parent === null);
-
-  // Order pages by menuOrder
-  navPages.sort((a, b) => parseFloat(a.menuOrder) - parseFloat(b.menuOrder));
-
-  return navPages;
-}
-
-/**
- * mapPageData
- */
-
-export function mapPageData(page = {}) {
-  const data = {
-    ...page,
-    ...page.page,
-  };
-
-  if (data.featuredImage) {
-    data.featuredImage = data.featuredImage.node;
-  }
-
-  if (data.parent) {
-    data.parent = data.parent.node;
-  }
-
-  if (data.children) {
-    data.children = data.children.edges.map(({ node }) => node);
-  }
-
-  return data;
-}
-
-/**
- * getBreadcrumbsByUri
- */
-
-export function getBreadcrumbsByUri(uri, pages) {
+  const segments = uri.split('/').filter(Boolean);
   const breadcrumbs = [];
-  const uriSegments = uri.split('/').filter((segment) => segment !== '');
 
-  // We don't want to show the current page in the breadcrumbs, so pop off
-  // the last chunk before we start
-
-  uriSegments.pop();
-
-  // Work through each of the segments, popping off the last chunk and finding the related
-  // page to gather the metadata for the breadcrumbs
-
-  do {
-    const breadcrumb = pages.find((page) => page.uri === `/${uriSegments.join('/')}/`);
-
-    // If the breadcrumb is the active page, we want to pass udefined for the uri to
-    // avoid the breadcrumbs being rendered as a link, given it's the current page
-
-    if (breadcrumb) {
+  let currentPath = '';
+  segments.forEach((segment) => {
+    currentPath += `/${segment}/`;
+    const page = pages.find((p) => p.uri === currentPath);
+    if (page) {
       breadcrumbs.push({
-        id: breadcrumb.id,
-        title: breadcrumb.title,
-        uri: breadcrumb.uri,
+        title: page.title,
+        uri: page.uri,
       });
     }
-
-    uriSegments.pop();
-  } while (uriSegments.length > 0);
-
-  // When working through the segments, we're doing so from the lowest child to the parent
-  // which means the parent will be at the end of the array. We need to reverse to show
-  // the correct order for breadcrumbs
-
-  breadcrumbs.reverse();
+  });
 
   return breadcrumbs;
 }
